Update README.md
This will be lab 2 as we learn the professional steps for creating a conceptual model of a desired information system. 

**Add a unique mandatory  additional Domain value type  to Athlete--

"YourNameHistory"

make the data type varchar 200 

*********************************

 NORMA lab 2 , starts on pg 63

Give me a report with a screen shot of your ORM diagram including:
Shot of the generated relational schema

Report should include lessons learned 

What are the primary keys

What are the foreign keys

Summary: What did you learn new?